docker-demo-ebpf/requirements.txt
flask>=3.0.0
requests>=2.31.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-api>=1.24.0
prometheus-client>=0.20.0
python-snappy>=0.7.0
protobuf>=5.0.0
